Separating new Pytest scripts and old ones
Can use this commit to see all the way normal tests were converted to pytest ones by looking at the modified files in scripts/
