+++ old_scripts/test_system.py
@@ -6,11 +6,6 @@ import time
 from rules_constants import *
 from vehicle_constants import *
 
-import pytest_check as check
-import pytest
-
-
-# ---------------------------------------------------------------------------- #
 AMS_STAT_OKAY = 1
 AMS_STAT_TRIP = 0
 AMS_CTRL_OKAY = 1
@@ -53,33 +48,10 @@ def cycle_power():
     time.sleep(CYCLE_POWER_OFF_DELAY)
     power.state = 0
     time.sleep(CYCLE_POWER_ON_DELAY)
-# ---------------------------------------------------------------------------- #
 
-
-# ---------------------------------------------------------------------------- #
-@pytest.fixture(scope="session")
-def hil():
-    global power
-
-    hil_instance = HIL()
-
-    hil_instance.load_config("config_charger.json")
-    hil_instance.load_pin_map("per_24_net_map.csv", "stm32f407_pin_map.csv")
-    
-    # hil_instance.init_can()
-    
-    power = hil_instance.dout("RearTester", "RLY1")
-    
-    yield hil_instance
-    
-    hil_instance.shutdown() 
-# ---------------------------------------------------------------------------- #
-
-
-# ---------------------------------------------------------------------------- #
 def test_precharge(hil):
     # Begin the test
-    # hil.start_test(test_precharge.__name__)
+    hil.start_test(test_precharge.__name__)
 
     # Outputs
     v_bat = hil.aout("Main_Module", "VBatt")
@@ -107,17 +79,13 @@ def test_precharge(hil):
     # Initial State
     reset_pchg(v_bat, v_mc)
     time.sleep(2.5)
-    # hil.check(pchg_cmplt.state == 0, "Precharge not complete on startup")
-    # hil.check(not_pchg_cmplt_delayed.state == 1, "Not precharge complete delayed high on startup")
-    check.equal(pchg_cmplt.state, 0, "Precharge not complete on startup")
-    check.equal(not_pchg_cmplt_delayed.state, 1, "Not precharge complete delayed high on startup")
+    hil.check(pchg_cmplt.state == 0, "Precharge not complete on startup")
+    hil.check(not_pchg_cmplt_delayed.state == 1, "Not precharge complete delayed high on startup")
     # Check delay
     v_mc.state = tiff_hv_to_lv(ACCUM_NOM_V)
     t = utils.measure_trip_time(not_pchg_cmplt_delayed, PCHG_COMPLETE_DELAY_S*3, is_falling=True)
-    # hil.check(not_pchg_cmplt_delayed.state == 0, "Precharge complete delayed")
-    # hil.check_within(t, PCHG_COMPLETE_DELAY_S, 0.25, f"Precharge complete delay of {t:.3}s close to expected {PCHG_COMPLETE_DELAY_S}s")
-    check.equal(not_pchg_cmplt_delayed.state, 0, "Precharge complete delayed")
-    check.almost_equal(t, PCHG_COMPLETE_DELAY_S, abs=0.25, rel=0.0, msg=f"Precharge complete delay of {t:.3}s close to expected {PCHG_COMPLETE_DELAY_S}s")
+    hil.check(not_pchg_cmplt_delayed.state == 0, "Precharge complete delayed")
+    hil.check_within(t, PCHG_COMPLETE_DELAY_S, 0.25, f"Precharge complete delay of {t:.3}s close to expected {PCHG_COMPLETE_DELAY_S}s")
 
     # Find threshold at nominal pack voltage
     for v in [ACCUM_MIN_V, ACCUM_NOM_V, ACCUM_MAX_V]:
@@ -127,8 +95,7 @@ def test_precharge(hil):
         v_mc.state = tiff_hv_to_lv(v*0.8)
         #time.sleep(0.01)
         time.sleep(0.5)
-        # hil.check(pchg_cmplt.state == 0, "Precharge Complete Low at Initial State")
-        check.equal(pchg_cmplt.state, 0, "Precharge Complete Low at Initial State")
+        hil.check(pchg_cmplt.state == 0, "Precharge Complete Low at Initial State")
 
         start = tiff_hv_to_lv(v*0.8)
         stop  = tiff_hv_to_lv(v)
@@ -137,52 +104,41 @@ def test_precharge(hil):
                                            pchg_cmplt, is_falling=0)
         thresh_hv = tiff_lv_to_hv(thresh)
         print(f"Precharge triggered at {thresh_hv / v * 100:.4}% ({thresh_hv:.5}V) of vbat={v}.")
-        # hil.check_within(thresh_hv / v, R_PCHG_V_BAT_THRESH, 0.03, f"Precharge threshold of {R_PCHG_V_BAT_THRESH*100}% at vbat = {v}V")
-        check.almost_equal(thresh_hv / v, R_PCHG_V_BAT_THRESH, abs=0.03, rel=0.0, msg=f"Precharge threshold of {R_PCHG_V_BAT_THRESH*100}% at vbat = {v}V")
+        hil.check_within(thresh_hv / v, R_PCHG_V_BAT_THRESH, 0.03, f"Precharge threshold of {R_PCHG_V_BAT_THRESH*100}% at vbat = {v}V")
         v_mc.state = tiff_hv_to_lv(v)
         #time.sleep(0.25)
         time.sleep(8)
-        # hil.check(pchg_cmplt.state == 1, f"Precharge completed at vbat = {v}V")
-        check.equal(pchg_cmplt.state, 1, f"Precharge completed at vbat = {v}V")
+        hil.check(pchg_cmplt.state == 1, f"Precharge completed at vbat = {v}V")
 
     
     # Floating conditions (check never precharge complete)
     reset_pchg(v_bat, v_mc)
     v_bat.hiZ()
     v_mc.state = tiff_hv_to_lv(0)
-    # hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat float, v_mc 0V")
-    check.equal(pchg_cmplt.state, 0, "Precharge not complete on v_bat float, v_mc 0V")
+    hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat float, v_mc 0V")
     v_mc.state = tiff_hv_to_lv(ACCUM_MAX_V)
-    # hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat float, v_mc max V")
-    check.equal(pchg_cmplt.state, 0, "Precharge not complete on v_bat float, v_mc max V")
+    hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat float, v_mc max V")
 
     reset_pchg(v_bat, v_mc)
     v_mc.hiZ()
     v_bat.state = tiff_hv_to_lv(ACCUM_MIN_V)
-    # hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat min, v_mc float")
-    check.equal(pchg_cmplt.state, 0, "Precharge not complete on v_bat min, v_mc float")
+    hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat min, v_mc float")
     v_bat.state = tiff_hv_to_lv(ACCUM_MAX_V)
-    # hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat max, v_mc float")
-    check.equal(pchg_cmplt.state, 0, "Precharge not complete on v_bat max, v_mc float")
+    hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat max, v_mc float")
 
     reset_pchg(v_bat, v_mc)
     v_bat.hiZ()
     v_mc.hiZ()
-    # hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat float, v_mc float")
-    check.equal(pchg_cmplt.state, 0, "Precharge not complete on v_bat float, v_mc float")
+    hil.check(pchg_cmplt.state == 0, "Precharge not complete on v_bat float, v_mc float")
 
     # TODO: software precharge validity checks (make precharge take forever)
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 BRK_SWEEP_DELAY = 0.1
 BSPD_DASH_ON_TIME = 0
-
 def test_bspd(hil):
     # Begin the test
-    # hil.start_test(test_bspd.__name__)
+    hil.start_test(test_bspd.__name__)
 
     # Outputs
     brk1    = hil.aout("Dashboard",   "BRK1_RAW")
@@ -207,29 +163,23 @@ def test_bspd(hil):
     # Brake threshold check
     brk1.state = BRK_1_REST_V
     brk2.state = BRK_2_REST_V
-    # hil.check(brk_stat_tap.state == 0, "Brake stat starts low")
-    check.equal(brk_stat_tap.state, 0, "Brake stat starts low")
+    hil.check(brk_stat_tap.state == 0, "Brake stat starts low")
     brk1.state = BRK_1_THRESH_V
     time.sleep(0.1)
-    # hil.check(brk_stat_tap.state == 1, "Brake stat goes high at brk 1 thresh")
-    check.equal(brk_stat_tap.state, 1, "Brake stat goes high at brk 1 thresh")
+    hil.check(brk_stat_tap.state == 1, "Brake stat goes high at brk 1 thresh")
     brk1.state = BRK_1_REST_V
-    # hil.check(brk_stat_tap.state == 0, "Brake stat starts low")
-    check.equal(brk_stat_tap.state, 0, "Brake stat starts low")
+    hil.check(brk_stat_tap.state == 0, "Brake stat starts low")
     brk2.state = BRK_2_THRESH_V
     time.sleep(0.1)
-    # hil.check(brk_stat_tap.state == 1, "Brake stat goes high at brk 2 thresh")
-    check.equal(brk_stat_tap.state, 1, "Brake stat goes high at brk 2 thresh")
+    hil.check(brk_stat_tap.state == 1, "Brake stat goes high at brk 2 thresh")
     brk1.state = BRK_1_THRESH_V
-    # hil.check(brk_stat_tap.state == 1, "Brake stat stays high for both brakes")
-    check.equal(brk_stat_tap.state, 1, "Brake stat stays high for both brakes")
+    hil.check(brk_stat_tap.state == 1, "Brake stat stays high for both brakes")
 
     # Brake threshold scan
     brk1.state = BRK_MIN_OUT_V
     brk2.state = BRK_MIN_OUT_V
     time.sleep(0.1)
-    # hil.check(brk_stat_tap.state == 0, "Brake Stat Starts Low Brk 1")
-    check.equal(brk_stat_tap.state, 0, "Brake Stat Starts Low Brk 1")
+    hil.check(brk_stat_tap.state == 0, "Brake Stat Starts Low Brk 1")
 
     start = BRK_MIN_OUT_V
     stop  = BRK_MAX_OUT_V
@@ -239,108 +189,72 @@ def test_bspd(hil):
                                        BRK_SWEEP_DELAY,
                                        brk_stat_tap, is_falling=False)
     print(f"Brake 1 braking threshold: {thresh}")
-    # hil.check_within(thresh, BRK_1_THRESH_V, 0.2, "Brake 1 trip voltage")
-    # hil.check(brk_stat_tap.state == 1, "Brake Stat Tripped for Brk 1")
-    check.almost_equal(thresh, BRK_1_THRESH_V, abs=0.2, rel=0.0, msg="Brake 1 trip voltage")
-    check.equal(brk_stat_tap.state, 1, "Brake Stat Tripped for Brk 1")
+    hil.check_within(thresh, BRK_1_THRESH_V, 0.2, "Brake 1 trip voltage")
+    hil.check(brk_stat_tap.state == 1, "Brake Stat Tripped for Brk 1")
 
     brk1.state = BRK_MIN_OUT_V
     brk2.state = BRK_MIN_OUT_V
-    # hil.check(brk_stat_tap.state == 0, "Brake Stat Starts Low Brk 2")
-    check.equal(brk_stat_tap.state, 0, "Brake Stat Starts Low Brk 2")
+    hil.check(brk_stat_tap.state == 0, "Brake Stat Starts Low Brk 2")
     thresh = utils.measure_trip_thresh(brk2, start, stop, step,
                                        BRK_SWEEP_DELAY,
                                        brk_stat_tap, is_falling=False)
     print(f"Brake 2 braking threshold: {thresh}")
-    # hil.check_within(thresh, BRK_2_THRESH_V, 0.2, "Brake 2 trip voltage")
-    # hil.check(brk_stat_tap.state == 1, "Brake Stat Tripped for Brk 2")
-    check.almost_equal(thresh, BRK_2_THRESH_V, abs=0.2, rel=0.0, msg="Brake 2 trip voltage")
-    check.equal(brk_stat_tap.state, 1, "Brake Stat Tripped for Brk 2")
+    hil.check_within(thresh, BRK_2_THRESH_V, 0.2, "Brake 2 trip voltage")
+    hil.check(brk_stat_tap.state == 1, "Brake Stat Tripped for Brk 2")
 
     # Brake Fail scan
     brk1.state = BRK_1_REST_V
     brk2.state = BRK_2_REST_V
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail Check 1 Starts 0")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail Check 1 Starts 0")
-
+    hil.check(brk_fail_tap.state == 0, "Brake Fail Check 1 Starts 0")
     brk1.state = 0.0 # Force 0
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 1 Short GND")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail Brk 1 Short GND")
-
+    hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 1 Short GND")
     brk1.state = BRK_1_REST_V
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail Check 2 Starts 0")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail Check 2 Starts 0")
-    
+    hil.check(brk_fail_tap.state == 0, "Brake Fail Check 2 Starts 0")
     brk2.state = 0.0 # Force 0
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 2 Short GND")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail Brk 2 Short GND")
-    
+    hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 2 Short GND")
     brk2.state = BRK_2_REST_V
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail Check 3 Starts 0")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail Check 3 Starts 0")
-    
+    hil.check(brk_fail_tap.state == 0, "Brake Fail Check 3 Starts 0")
     brk1.state = 5.0 # Short VCC
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 1 Short VCC")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail Brk 1 Short VCC")
-    
+    hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 1 Short VCC")
     brk1.state = BRK_1_REST_V
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail Check 4 Starts 0")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail Check 4 Starts 0")
-    
+    hil.check(brk_fail_tap.state == 0, "Brake Fail Check 4 Starts 0")
     brk2.state = 5.0 # Short VCC
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 2 Short VCC")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail Brk 2 Short VCC")
-    
+    hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 2 Short VCC")
     brk2.state = BRK_2_REST_V
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail Check 5 Starts 0")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail Check 5 Starts 0")
-    
+    hil.check(brk_fail_tap.state == 0, "Brake Fail Check 5 Starts 0")
     brk1.hiZ()
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 1 Hi-Z")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail Brk 1 Hi-Z")
-    
+    hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 1 Hi-Z")
     brk1.state = BRK_1_REST_V
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail Check 6 Starts 0")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail Check 6 Starts 0")
-    
+    hil.check(brk_fail_tap.state == 0, "Brake Fail Check 6 Starts 0")
     brk2.hiZ()
     time.sleep(0.1)
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 2 Hi-Z")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail Brk 2 Hi-Z")
-    
+    hil.check(brk_fail_tap.state == 1, "Brake Fail Brk 2 Hi-Z")
     brk2.state = BRK_2_REST_V
 
     # Brake Fail
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME)
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
-    
+    hil.check(bspd_ctrl.state == 1, "Power On")
     brk1.state = 0.0
     t = utils.measure_trip_time(bspd_ctrl, 5.0, is_falling=True)
-    # hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Brake Fail")
-    # hil.check(brk_fail_tap.state == 1, "Brake Fail went high")
-    check.less(t, R_BSPD_MAX_TRIP_TIME_S, "Brake Fail")
-    check.equal(brk_fail_tap.state, 1, "Brake Fail went high")
-    
+    hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Brake Fail")
+    hil.check(brk_fail_tap.state == 1, "Brake Fail went high")
     brk1.state = BRK_1_REST_V
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(brk_fail_tap.state == 0, "Brake Fail returned low")
-    # hil.check(bspd_ctrl.state == 0, "Brake Fail Stays Latched")
-    check.equal(brk_fail_tap.state, 0, "Brake Fail returned low")
-    check.equal(bspd_ctrl.state, 0, "Brake Fail Stays Latched")
+    hil.check(brk_fail_tap.state == 0, "Brake Fail returned low")
+    hil.check(bspd_ctrl.state == 0, "Brake Fail Stays Latched")
 
     # Brake Fail on Power On
     reset_bspd(brk1, brk2, c_sense)
@@ -352,8 +266,7 @@ def test_bspd(hil):
     power.state = 0
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
     time.sleep(BSPD_DASH_ON_TIME) # NOTE: test can't check for the trip time
-    # hil.check(bspd_ctrl.state == 0, "Power On Brake Fail")
-    check.equal(bspd_ctrl.state, 0, "Power On Brake Fail")
+    hil.check(bspd_ctrl.state == 0, "Power On Brake Fail")
 
     # Current no brake
     reset_bspd(brk1, brk2, c_sense)
@@ -367,59 +280,47 @@ def test_bspd(hil):
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
     # time.sleep(100)
     time.sleep(3)
-    # hil.check(bspd_ctrl.state == 1, "Current no brake")
-    check.equal(bspd_ctrl.state, 1, "Current no brake")
+    hil.check(bspd_ctrl.state == 1, "Current no brake")
 
     # Current Sense Short to Ground
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME)
-    # hil.check(bspd_ctrl.state == 1, "Power On"
-    check.equal(bspd_ctrl.state, 1, "Power On")
-
+    hil.check(bspd_ctrl.state == 1, "Power On")
     c_sense.state = 0.0
     t = utils.measure_trip_time(bspd_ctrl, 5.0, is_falling=True)
-    # hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Current short to ground")
-    check.less(t, R_BSPD_MAX_TRIP_TIME_S, "Current short to ground")
+    hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Current short to ground")
     set_bspd_current(c_sense, 0.0)
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 0, "Current short to ground stays latched")
-    check.equal(bspd_ctrl.state, 0, "Current short to ground stays latched")
+    hil.check(bspd_ctrl.state == 0, "Current short to ground stays latched")
 
     # Current Sense Short to 5V
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME*1.2)
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
+    hil.check(bspd_ctrl.state == 1, "Power On")
     c_sense.state = ABOX_DHAB_CH1_DIV.div(5.0)
     t = utils.measure_trip_time(bspd_ctrl, 5.0, is_falling=True)
-    # hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Current short to 5V")
-    check.less(t, R_BSPD_MAX_TRIP_TIME_S, "Current short to 5V")
+    hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Current short to 5V")
     set_bspd_current(c_sense, 0.0)
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 0, "Current short to 5V stays latched")
-    check.equal(bspd_ctrl.state, 0, "Current short to 5V stays latched")
+    hil.check(bspd_ctrl.state == 0, "Current short to 5V stays latched")
 
     # Braking
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME)
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
+    hil.check(bspd_ctrl.state == 1, "Power On")
     brk1.state = BRK_1_THRESH_V
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 1, "Brake no current")
-    # hil.check(brk_stat_tap.state == 1, "Brake stat went high")
-    check.equal(bspd_ctrl.state, 1, "Brake no current")
-    check.equal(brk_stat_tap.state, 1, "Brake stat went high")
+    hil.check(bspd_ctrl.state == 1, "Brake no current")
+    hil.check(brk_stat_tap.state == 1, "Brake stat went high")
 
     # Lowest current required to trip at
     min_trip_current = R_BSPD_POWER_THRESH_W / ACCUM_MAX_V
     set_bspd_current(c_sense, min_trip_current)
     t = utils.measure_trip_time(bspd_ctrl, 10.0, is_falling=True)
-    # hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Braking with current")
-    check.less(t, R_BSPD_MAX_TRIP_TIME_S, "Braking with current")
+    hil.check(t < R_BSPD_MAX_TRIP_TIME_S, "Braking with current")
 
     # Measure braking with current threshold
     reset_bspd(brk1, brk2, c_sense)
@@ -427,8 +328,7 @@ def test_bspd(hil):
     time.sleep(BSPD_DASH_ON_TIME)
     brk1.state = BRK_1_THRESH_V
 
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
+    hil.check(bspd_ctrl.state == 1, "Power On")
     start = ABOX_DHAB_CH1_DIV.div(dhab_ch1_a_to_v(0.0))
     stop  = ABOX_DHAB_CH1_DIV.div(dhab_ch1_a_to_v(DHAB_S124_CH1_MAX_A))
     step  = 0.1
@@ -437,23 +337,16 @@ def test_bspd(hil):
                                        bspd_ctrl, is_falling=True)
     thresh_amps = dhab_ch1_v_to_a(ABOX_DHAB_CH1_DIV.reverse(thresh))
     print(f"Current while braking threshold: {thresh}V = {thresh_amps}A")
-    # hil.check_within(thresh, ABOX_DHAB_CH1_DIV.div(dhab_ch1_a_to_v(min_trip_current)), 0.1, "Current while braking threshold")
-    check.almost_equal(
-        thresh, ABOX_DHAB_CH1_DIV.div(dhab_ch1_a_to_v(min_trip_current)),
-        abs=0.1, rel=0.0,
-        msg="Current while braking threshold"
-    )
+    hil.check_within(thresh, ABOX_DHAB_CH1_DIV.div(dhab_ch1_a_to_v(min_trip_current)), 0.1, "Current while braking threshold")
 
     # Determine the current sense short to gnd threshold
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME)
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
+    hil.check(bspd_ctrl.state == 1, "Power On")
     set_bspd_current(c_sense, DHAB_S124_CH1_MIN_A)
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 1, "Min output current okay")
-    check.equal(bspd_ctrl.state, 1, "Min output current okay")
+    hil.check(bspd_ctrl.state == 1, "Min output current okay")
     start = ABOX_DHAB_CH1_DIV.div(DHAB_S124_MIN_OUT_V)
     stop = 0.0
     step = -0.1
@@ -461,21 +354,18 @@ def test_bspd(hil):
                                        R_BSPD_MAX_TRIP_TIME_S,
                                        bspd_ctrl, is_falling=True)
     print(f"Short to ground threshold: {thresh}V")
-    # hil.check(stop < (thresh) < start, "Current short to ground threshold")
-    check.between(thresh, stop, start, "Current short to ground threshold")
+    hil.check(stop < (thresh) < start, "Current short to ground threshold")
 
     # Determine the current sense short to 5V threshold
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME)
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
+    hil.check(bspd_ctrl.state == 1, "Power On")
     time.sleep(2)
 
     set_bspd_current(c_sense, DHAB_S124_CH1_MAX_A)
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 1, "Max output current okay")
-    check.equal(bspd_ctrl.state, 1, "Max output current okay")
+    hil.check(bspd_ctrl.state == 1, "Max output current okay")
 
     start = ABOX_DHAB_CH1_DIV.div(DHAB_S124_MAX_OUT_V)
     stop = ABOX_DHAB_CH1_DIV.div(5.0)
@@ -483,31 +373,26 @@ def test_bspd(hil):
 
     c_sense.state = start
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 1, "Max output voltage okay")
-    check.equal(bspd_ctrl.state, 1, "Max output voltage okay")
+    hil.check(bspd_ctrl.state == 1, "Max output voltage okay")
     input("enter to continue")
 
     thresh = utils.measure_trip_thresh(c_sense, start, stop, step,
                                        R_BSPD_MAX_TRIP_TIME_S,
                                        bspd_ctrl, is_falling=True)
     print(f"Short to 5V threshold: {thresh}V")
-    # hil.check(bspd_ctrl.state == 0, "Short to 5V trips")
-    check.equal(bspd_ctrl.state, 0, "Short to 5V trips")
+    hil.check(bspd_ctrl.state == 0, "Short to 5V trips")
     print(stop)
     print(start)
-    # hil.check(start < (thresh) <= stop, "Current short to 5V threshold")
-    check.between(thresh, start, stop, "Current short to 5V threshold", le=True)
+    hil.check(start < (thresh) <= stop, "Current short to 5V threshold")
 
     # Floating current
     reset_bspd(brk1, brk2, c_sense)
     cycle_power()
     time.sleep(BSPD_DASH_ON_TIME)
-    # hil.check(bspd_ctrl.state == 1, "Power On")
-    check.equal(bspd_ctrl.state, 1, "Power On")
+    hil.check(bspd_ctrl.state == 1, "Power On")
     c_sense.hiZ()
     time.sleep(R_BSPD_MAX_TRIP_TIME_S)
-    # hil.check(bspd_ctrl.state == 0, "Floating current")
-    check.equal(bspd_ctrl.state, 0, "Floating current")
+    hil.check(bspd_ctrl.state == 0, "Floating current")
 
     # Floating brake_fail
     # Can't test this at system level!
@@ -516,11 +401,9 @@ def test_bspd(hil):
     # Can't test this at system level!
 
     # End the test
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
 
-# ---------------------------------------------------------------------------- #
 IMD_RC_MIN_TRIP_TIME_S = IMD_STARTUP_TIME_S 
 IMD_RC_MAX_TRIP_TIME_S = R_IMD_MAX_TRIP_TIME_S - IMD_MEASURE_TIME_S
 IMD_CTRL_OKAY = 1
@@ -528,7 +411,7 @@ IMD_CTRL_TRIP = 0
 
 def test_imd(hil):
     # Begin the test
-    # hil.start_test(test_imd.__name__)
+    hil.start_test(test_imd.__name__)
 
     # Outputs
     imd_stat  = hil.dout("Main_Module", "IMD_Status")
@@ -549,47 +432,37 @@ def test_imd(hil):
     # IMD Fault
     reset_imd(imd_stat)
     cycle_power()
-    # hil.check(imd_ctrl.state == IMD_CTRL_OKAY, "Power On")
-    check.equal(imd_ctrl.state, IMD_CTRL_OKAY, "Power On")
+    hil.check(imd_ctrl.state == IMD_CTRL_OKAY, "Power On")
     time.sleep(1)
     imd_stat.state = IMD_STAT_TRIP
     t = utils.measure_trip_time(imd_ctrl, R_IMD_MAX_TRIP_TIME_S, is_falling=True)
     print(f"Target trip time: [{IMD_RC_MIN_TRIP_TIME_S}, {IMD_RC_MAX_TRIP_TIME_S}]")
-    # hil.check(IMD_RC_MIN_TRIP_TIME_S < t < IMD_RC_MAX_TRIP_TIME_S, "IMD Trip Time")
-    # hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Trip")
-    check.between(t, IMD_RC_MIN_TRIP_TIME_S, IMD_RC_MAX_TRIP_TIME_S, "IMD Trip Time")
-    check.equal(imd_ctrl.state, IMD_CTRL_TRIP, "IMD Trip")
+    hil.check(IMD_RC_MIN_TRIP_TIME_S < t < IMD_RC_MAX_TRIP_TIME_S, "IMD Trip Time")
+    hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Trip")
     imd_stat.state = IMD_STAT_OKAY
     time.sleep(IMD_RC_MAX_TRIP_TIME_S * 1.1)
-    # hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Fault Stays Latched")
-    check.equal(imd_ctrl.state, IMD_CTRL_TRIP, "IMD Fault Stays Latched")
+    hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Fault Stays Latched")
 
     # IMD Fault on Power On
     reset_imd(imd_stat)
     imd_stat.state = IMD_STAT_TRIP
     cycle_power()
     time.sleep(IMD_RC_MAX_TRIP_TIME_S)
-    # hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Fault Power On")
-    check.equal(imd_ctrl.state, IMD_CTRL_TRIP, "IMD Fault Power On")
+    hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Fault Power On")
 
     # IMD Floating
     reset_imd(imd_stat)
     imd_stat.hiZ()
     cycle_power()
     t = utils.measure_trip_time(imd_ctrl, R_IMD_MAX_TRIP_TIME_S, is_falling=True)
-    # hil.check(t < R_IMD_MAX_TRIP_TIME_S, "IMD Floating Trip Time")
-    # hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Floating Trip")
-    check.less(t, R_IMD_MAX_TRIP_TIME_S, "IMD Floating Trip Time")
-    check.equal(imd_ctrl.state, IMD_CTRL_TRIP, "IMD Floating Trip")
+    hil.check(t < R_IMD_MAX_TRIP_TIME_S, "IMD Floating Trip Time")
+    hil.check(imd_ctrl.state == IMD_CTRL_TRIP, "IMD Floating Trip")
     
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 def test_ams(hil):
     # Begin the test
-    # hil.start_test(test_ams.__name__)
+    hil.start_test(test_ams.__name__)
 
     # Outputs
     ams_stat  = hil.dout("Main_Module", "BMS-Status-Main")
@@ -610,43 +483,33 @@ def test_ams(hil):
     # AMS Fault
     reset_ams(ams_stat)
     cycle_power()
-    # hil.check(ams_ctrl.state == AMS_CTRL_OKAY, "Power On")
-    check.equal(ams_ctrl.state, AMS_CTRL_OKAY, "Power On")
+    hil.check(ams_ctrl.state == AMS_CTRL_OKAY, "Power On")
     time.sleep(1)
     ams_stat.state = AMS_STAT_TRIP
     t = utils.measure_trip_time(ams_ctrl, AMS_MAX_TRIP_DELAY_S * 2, is_falling=True)
-    # hil.check(0 < t < AMS_MAX_TRIP_DELAY_S, "AMS Trip Time")
-    # hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Trip")
-    check.between(t, 0, AMS_MAX_TRIP_DELAY_S, "AMS Trip Time")
-    check.equal(ams_ctrl.state, AMS_CTRL_TRIP, "AMS Trip")
+    hil.check(0 < t < AMS_MAX_TRIP_DELAY_S, "AMS Trip Time")
+    hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Trip")
     ams_stat.state = AMS_STAT_OKAY
     time.sleep(AMS_MAX_TRIP_DELAY_S * 1.1)
-    # hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Fault Stays Latched")
-    check.equal(ams_ctrl.state, AMS_CTRL_TRIP, "AMS Fault Stays Latched")
+    hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Fault Stays Latched")
 
     # AMS Fault on Power On
     reset_ams(ams_stat)
     ams_stat.state = AMS_STAT_TRIP
     cycle_power()
     time.sleep(AMS_MAX_TRIP_DELAY_S)
-    # hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Fault Power On")
-    check.equal(ams_ctrl.state, AMS_CTRL_TRIP, "AMS Fault Power On")
+    hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Fault Power On")
 
     # AMS Floating
     reset_ams(ams_stat)
     ams_stat.hiZ()
     cycle_power()
     t = utils.measure_trip_time(ams_ctrl, AMS_MAX_TRIP_DELAY_S * 2, is_falling=True)
-    # hil.check(0 <= t < AMS_MAX_TRIP_DELAY_S, "AMS Floating Trip Time")
-    # hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Floating Trip")
-    check.between(t, 0, AMS_MAX_TRIP_DELAY_S, "AMS Floating Trip Time")
-    check.equal(ams_ctrl.state, AMS_CTRL_TRIP, "AMS Floating Trip")
+    hil.check(0 <= t < AMS_MAX_TRIP_DELAY_S, "AMS Floating Trip Time")
+    hil.check(ams_ctrl.state == AMS_CTRL_TRIP, "AMS Floating Trip")
     
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 def tsal_is_red():
     while 1:
         i = input("Is TSAL Green (g) or Red (r): ")
@@ -673,25 +536,21 @@ def test_tsal(hil):
     time.sleep(0.2)
     # No need to power cycle
 
-    # hil.check(tsal_is_red() == False, "LVAL on at v_mc = 0")
-    check.equal(tsal_is_red(), False, "LVAL on at v_mc = 0")
+    hil.check(tsal_is_red() == False, "LVAL on at v_mc = 0")
     #hil.check(tsal.state == 0, "TSAL off at v_mc = 0")
 
     time.sleep(5)
-    # hil.check(tsal_is_red() == False, "LVAL stays on")
-    check.equal(tsal_is_red(), False, "LVAL stays on")
+    hil.check(tsal_is_red() == False, "LVAL stays on")
 
     v_mc.state = tiff_hv_to_lv(ACCUM_MIN_V)
     time.sleep(0.1)
 
     #hil.check(lval.state == 0, f"LVAL off at {ACCUM_MIN_V:.4} V")
-    # hil.check(tsal_is_red() == True, f"TSAL on at {ACCUM_MIN_V:.4} V")
-    check.equal(tsal_is_red(), True, f"TSAL on at {ACCUM_MIN_V:.4} V")
+    hil.check(tsal_is_red() == True, f"TSAL on at {ACCUM_MIN_V:.4} V")
 
     reset_tsal(v_mc)
     time.sleep(0.2)
-    # hil.check(tsal_is_red() == False, f"LVAL turns back on")
-    check.equal(tsal_is_red(), False, f"LVAL turns back on")
+    hil.check(tsal_is_red() == False, f"LVAL turns back on")
 
     start = tiff_hv_to_lv(50)
     stop  = tiff_hv_to_lv(R_TSAL_HV_V * 1.5)
@@ -715,36 +574,27 @@ def test_tsal(hil):
     if (not tripped):
         utils.log_warning(f"TSAL did not trip at stop of {stop}.")
         thresh = stop
-    # hil.check(tripped, "TSAL tripped")
-    check.is_true(tripped, "TSAL tripped")
+    hil.check(tripped, "TSAL tripped")
     
     thresh = tiff_lv_to_hv(thresh)
     print(f"TSAL on at {thresh:.4} V")
     hil.check_within(thresh, R_TSAL_HV_V, 4, f"TSAL trips at {R_TSAL_HV_V:.4} +-4")
-    check.almost_equal(thresh, R_TSAL_HV_V, abs=4, rel=0.0, msg=f"TSAL trips at {R_TSAL_HV_V:.4} +-4")
 
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 def test_sdc(hil):
     ''' Check that every node in the sdc trips '''
     # Begin the test
-    # hil.start_test(test_sdc.__name__)
+    hil.start_test(test_sdc.__name__)
 
     # Outputs
 
     # Inputs
 
-    # hil.check(False, "TODO")
-    check.is_true(False, "TODO")
+    hil.check(0, "TODO")
 
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 def is_buzzer_on():
     while 1:
         i = input("Is Buzzer On (y) or No (n): ")
@@ -756,7 +606,7 @@ def is_buzzer_on():
 
 def test_buzzer(hil):
     # Begin the test
-    # hil.start_test(test_buzzer.__name__)
+    hil.start_test(test_buzzer.__name__)
 
     # Outputs
     buzzer_ctrl = hil.daq_var("Main_Module", "daq_buzzer") 
@@ -766,29 +616,21 @@ def test_buzzer(hil):
 
     buzzer_ctrl.state = 0
     time.sleep(0.02)
-    # hil.check(buzzer_stat.state == 0, "Buzzer Off")
-    check.equal(buzzer_stat.state, 0, "Buzzer Off")
+    hil.check(buzzer_stat.state == 0, "Buzzer Off")
 
     buzzer_ctrl.state = 1
     print(buzzer_ctrl.state)
     time.sleep(0.02)
-    # hil.check(buzzer_stat.state == 1, "Buzzer On")
-    # hil.check(is_buzzer_on() == True, "Buzzer Making Noise")
-    check.equal(buzzer_stat.state, 1, "Buzzer On")
-    check.is_true(is_buzzer_on(), "Buzzer Making Noise")
+    hil.check(buzzer_stat.state == 1, "Buzzer On")
+    hil.check(is_buzzer_on() == True, "Buzzer Making Noise")
 
     buzzer_ctrl.state = 0
     time.sleep(0.02)
-    # hil.check(buzzer_stat.state == 0, "Buzzer back Off")
-    # hil.check(is_buzzer_on() == False, "Buzzer Not Making Noise")
-    check.equal(buzzer_stat.state, 0, "Buzzer back Off")
-    check.is_false(is_buzzer_on(), "Buzzer Not Making Noise")
+    hil.check(buzzer_stat.state == 0, "Buzzer back Off")
+    hil.check(is_buzzer_on() == False, "Buzzer Not Making Noise")
 
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 def is_brake_light_on():
     while 1:
         i = input("Is Brake Light On (y) or No (n): ")
@@ -800,7 +642,7 @@ def is_brake_light_on():
 
 def test_brake_light(hil):
     # Begin the test
-    # hil.start_test(test_brake_light.__name__)
+    hil.start_test(test_brake_light.__name__)
 
     # Outputs
     brk_ctrl = hil.daq_var("Main_Module", "daq_brake") 
@@ -810,36 +652,28 @@ def test_brake_light(hil):
 
     brk_ctrl.state = 0
     time.sleep(0.02)
-    # hil.check(brk_ctrl.state == 0, "Brake Off")
-    check.equal(brk_ctrl.state, 0, "Brake Off")
+    hil.check(brk_ctrl.state == 0, "Brake Off")
 
     brk_ctrl.state = 1
     print(brk_ctrl.state)
     time.sleep(0.02)
-    # hil.check(brk_ctrl.state == 1, "Brake Light On")
-    # hil.check(is_brake_light_on() == True, "Brake Light is On")
-    check.equal(brk_ctrl.state, 1, "Brake Light On")
-    check.is_true(is_brake_light_on(), "Brake Light is On")
+    hil.check(brk_ctrl.state == 1, "Brake Light On")
+    hil.check(is_brake_light_on() == True, "Brake Light is On")
 
     brk_ctrl.state = 0
     time.sleep(0.02)
-    # hil.check(brk_ctrl.state == 0, "Brake Light back Off")
-    # hil.check(is_brake_light_on() == False, "Brake Light is Off")
-    check.equal(brk_ctrl.state, 0, "Brake Light back Off")
-    check.is_false(is_brake_light_on(), "Brake Light is Off")
+    hil.check(brk_ctrl.state == 0, "Brake Light back Off")
+    hil.check(is_brake_light_on() == False, "Brake Light is Off")
 
 
     # Can copy lot from bspd
     # Read the brake control mcu pin
     # Finally have user verify light actually turned on
 
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+    hil.end_test()
 
-
-# ---------------------------------------------------------------------------- #
 def test_light_tsal_buz(hil):
-    # hil.start_test(test_light_tsal_buz.__name__)
+    hil.start_test(test_light_tsal_buz.__name__)
 
     # Outputs
     brk_ctrl = hil.daq_var("Main_Module", "daq_brake") 
@@ -851,7 +685,30 @@ def test_light_tsal_buz(hil):
     brk_ctrl.state = 0
     buzzer_ctrl.state = 0
 
-    check.is_true(True, "TODO")
+    hil.end_test()
 
-    # hil.end_test()
-# ---------------------------------------------------------------------------- #
+
+if __name__ == "__main__":
+    hil = HIL()
+
+    hil.load_config("config_system_hil_attached.json")
+    hil.load_pin_map("per_24_net_map.csv", "stm32f407_pin_map.csv")
+
+    hil.init_can()
+
+    power = hil.dout("RearTester", "RLY1")
+
+    # Drive Critical Tests
+    # test_precharge(hil)
+    test_bspd(hil)
+    # test_imd(hil) # note: tsal needs to be tripped
+    # test_ams(hil)
+    # test_tsal(hil)
+    # test_sdc(hil)
+    # test_buzzer(hil)
+    # test_brake_light(hil)
+    # test_light_tsal_buz(hil)
+
+    # Peripheral Sensor Tests
+
+    hil.shutdown()
